Move StableDiffusion directory
