+++ keras_cv/models/__init__.py
@@ -43,7 +43,6 @@ from keras_cv.models.efficientnet_v2 import EfficientNetV2B3
 from keras_cv.models.efficientnet_v2 import EfficientNetV2L
 from keras_cv.models.efficientnet_v2 import EfficientNetV2M
 from keras_cv.models.efficientnet_v2 import EfficientNetV2S
-from keras_cv.models.generative.stable_diffusion import StableDiffusion
 from keras_cv.models.mlp_mixer import MLPMixerB16
 from keras_cv.models.mlp_mixer import MLPMixerB32
 from keras_cv.models.mlp_mixer import MLPMixerL16
@@ -59,4 +58,5 @@ from keras_cv.models.resnet_v1 import ResNet152
 from keras_cv.models.resnet_v2 import ResNet50V2
 from keras_cv.models.resnet_v2 import ResNet101V2
 from keras_cv.models.resnet_v2 import ResNet152V2
+from keras_cv.models.stable_diffusion import StableDiffusion
 from keras_cv.models.vgg19 import VGG19
