+++ keras_cv/models/stable_diffusion/README.md
