+++ keras_cv/models/stable_diffusion/__init__.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from keras_cv.models.generative.stable_diffusion.stable_diffusion import Decoder
-from keras_cv.models.generative.stable_diffusion.stable_diffusion import DiffusionModel
-from keras_cv.models.generative.stable_diffusion.stable_diffusion import ImageEncoder
-from keras_cv.models.generative.stable_diffusion.stable_diffusion import SimpleTokenizer
-from keras_cv.models.generative.stable_diffusion.stable_diffusion import StableDiffusion
-from keras_cv.models.generative.stable_diffusion.stable_diffusion import TextEncoder
+from keras_cv.models.stable_diffusion.decoder import Decoder
+from keras_cv.models.stable_diffusion.diffusion_model import DiffusionModel
+from keras_cv.models.stable_diffusion.image_encoder import ImageEncoder
+from keras_cv.models.stable_diffusion.clip_tokenizer import SimpleTokenizer
+from keras_cv.models.stable_diffusion.stable_diffusion import StableDiffusion
+from keras_cv.models.stable_diffusion.text_encoder import TextEncoder
