+++ keras_cv/models/stable_diffusion/__internal__/__init__.py
