+++ keras_cv/models/stable_diffusion/__internal__/layers/__init__.py
