+++ keras_cv/models/stable_diffusion/__internal__/layers/group_normalization.py
