+++ keras_cv/models/stable_diffusion/__internal__/layers/padded_conv2d.py
