+++ keras_cv/models/stable_diffusion/clip_tokenizer.py
