+++ keras_cv/models/stable_diffusion/constants.py
