+++ keras_cv/models/stable_diffusion/stable_diffusion_test.py
