+++ keras_cv/models/stable_diffusion/text_encoder.py
